+++ lib/src/custom/custom_toast.dart
@@ -10,13 +10,14 @@ class CustomToast extends BaseDialog {
     required OverlayEntry overlayEntry,
   }) : super(config: config, overlayEntry: overlayEntry);
 
-  List<void Function()> _toastList = [];
+  List<Future<void> Function()> _toastList = [];
 
   DateTime? _lastTime;
 
   Future<void> showToast({
     required Duration time,
     required bool antiShake,
+    required SmartToastStatus status,
     required Widget widget,
   }) async {
     // anti-shake
@@ -27,8 +28,22 @@ class CustomToast extends BaseDialog {
       _lastTime = now;
       if (isShake) return;
     }
-
     config.isExistToast = true;
+
+    // provider some toast display effect
+    if (status == SmartToastStatus.normal) {
+      _normalToast(time: time, widget: widget);
+    } else if (status == SmartToastStatus.last) {
+      _lastToast(time: time, widget: widget);
+    } else if (status == SmartToastStatus.firstAndLast) {
+      _firstAndLastToast(time: time, widget: widget);
+    }
+  }
+
+  Future<void> _normalToast({
+    required Duration time,
+    required Widget widget,
+  }) async {
     _toastList.add(() async {
       //handling special circumstances
       if (_toastList.length == 0) return;
@@ -51,10 +66,71 @@ class CustomToast extends BaseDialog {
       _toastList.removeAt(0);
       await dismiss();
 
-      if (_toastList.length != 0) _toastList[0]();
+      if (_toastList.length != 0) await _toastList[0]();
     });
 
-    if (_toastList.length == 1) _toastList[0]();
+    if (_toastList.length == 1) await _toastList[0]();
+  }
+
+  Future<void> _lastToast({
+    required Duration time,
+    required Widget widget,
+  }) async {
+    mainDialog.show(
+      alignment: Alignment.center,
+      maskColor: Colors.transparent,
+      maskWidget: null,
+      animationDuration: Duration(milliseconds: 200),
+      isLoading: true,
+      isUseAnimation: true,
+      isPenetrate: true,
+      clickBgDismiss: false,
+      widget: widget,
+      onDismiss: null,
+      onBgTap: () => dismiss(),
+    );
+    _toastList.add(() async {});
+    await Future.delayed(time);
+    if (_toastList.length == 1) {
+      await dismiss();
+    }
+    _toastList.removeLast();
+  }
+
+  Future<void> _firstAndLastToast({
+    required Duration time,
+    required Widget widget,
+  }) async {
+    _toastList.add(() async {
+      //handling special circumstances
+      if (_toastList.length == 0) return;
+
+      mainDialog.show(
+        alignment: Alignment.center,
+        maskColor: Colors.transparent,
+        maskWidget: null,
+        animationDuration: Duration(milliseconds: 200),
+        isLoading: true,
+        isUseAnimation: true,
+        isPenetrate: true,
+        clickBgDismiss: false,
+        widget: widget,
+        onDismiss: null,
+        onBgTap: () => dismiss(),
+      );
+      await Future.delayed(time);
+      //invoke next toast
+      _toastList.removeAt(0);
+      await dismiss();
+
+      if (_toastList.length != 0) await _toastList[0]();
+    });
+
+    if (_toastList.length == 1) await _toastList[0]();
+
+    if (_toastList.length > 2) {
+      _toastList.removeAt(1);
+    }
   }
 
   Future<void> dismiss() async {